requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
google-generativeai==0.3.2 